// ui/src/app/api/model/movie.ts
/**
 * Public Mollies Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocalMovie } from './local-movie';
import { TransmissionStatusCode } from './transmission-status-code';
import { MovieSource } from './movie-source';

export interface Movie {
  id?: number;
  metaSource?: string | null;
  imdbCode?: string | null;
  title?: string | null;
  language?: string | null;
  year?: number;
  rating?: number;
  description?: string | null;
  youTubeTrailerCode?: string | null;
  imageFilename?: string | null;
  genres?: Array<string> | null;
  movieSources?: Array<MovieSource> | null;
  localMovie?: LocalMovie;
  transmissionStatus?: TransmissionStatusCode;
}